Moved argument parsing logic out of mapper's main.cpp into a new file args.hpp so that multiple applications can use it.
